fix: 将db的异常抛出
