+++ src/repository/sqlite.py
@@ -2,6 +2,8 @@ from contextlib import asynccontextmanager
 from typing import AsyncGenerator
 import aiosqlite
 
+from .db import ActionExec, ActionIter
+
 
 @asynccontextmanager
 async def __transaction(conn: aiosqlite.Connection):
@@ -28,33 +30,6 @@ async def get_conn(db_name: str, use_transaction: bool = False) -> AsyncGenerato
     pass
 
 
-class ActionExec:
-    def __init__(self, sql: str, *args) -> None:
-        self.__sql = sql
-        self.__args = args
-        pass
-
-    async def __call__(self, cursor: aiosqlite.Cursor) -> int:
-        cursor_res = await cursor.execute(self.__sql, self.__args)
-        return cursor_res.rowcount
-    pass
-
-
-class ActionIter:
-    def __init__(self, sql: str, *args) -> None:
-        self.__sql = sql
-        self.__args = args
-        pass
-
-    async def __call__(self, cursor: aiosqlite.Cursor) -> AsyncGenerator[aiosqlite.Row, None]:
-        await cursor.execute(self.__sql, self.__args)
-        while (row := await cursor.fetchone()) is not None:
-            yield row
-            pass
-        pass
-    pass
-
-
 class ConnExecutor:
     def __init__(self, conn: aiosqlite.Connection) -> None:
         self.__conn = conn
@@ -64,7 +39,7 @@ class ConnExecutor:
         async with self.__conn.cursor() as cursor:
             return await coro(cursor)
 
-    async def iter(self, gen: ActionIter) -> AsyncGenerator[aiosqlite.Row, None]:
+    async def iter(self, gen: ActionIter) -> AsyncGenerator[dict, None]:
         async with self.__conn.cursor() as cursor:
             async for row in gen(cursor):
                 yield row
@@ -81,13 +56,7 @@ class SqliteManage:
 
     @asynccontextmanager
     async def __call__(self, use_transaction: bool = False) -> AsyncGenerator[ConnExecutor, None]:
-        try:
-            async with get_conn(self.__db_name, use_transaction) as conn:
-                yield ConnExecutor(conn)
-            pass
-        except Exception as e:
-            # TODO: 这里需要记录日志
-            print(e)
-            pass
+        async with get_conn(self.__db_name, use_transaction) as conn:
+            yield ConnExecutor(conn)
         pass
     pass
